replace dosbox cue parser with custom implementation

+++ src/dosbox-cdrom/cdrom.h
@@ -48,7 +48,7 @@ class CdImage
 private:
   struct Track
   {
-    int number = 0;
+    size_t number = 0;
     size_t start = 0;
     size_t length = 0;
     size_t skip = 0;
